Justin First Modification
Signup and signin interface
splash activity

+++ seniorProject - Copy/BaitBite_Chef/app/src/main/java/com/example/matsah/baitbite_chef/Model/Chef.java
@@ -8,8 +8,8 @@ public class Chef {
 
     //private String Account_ID;
     private String Email ="";
-    private int LocationX;
-    private int LocationY;
+    private double LocationX;
+    private double LocationY;
     private int Mobile_Verification_Code;
     private String Name;
     private String Phone_Number;
@@ -32,7 +32,7 @@ public class Chef {
     public Chef() {
     }
 
-    public Chef(String email, int locationX,int locationY, String name, String phone_number){
+    public Chef(String email, double locationX,double locationY, String name, String phone_number){
         this.Email = email;
         this.LocationX = locationX;
         this.LocationY = locationY;
@@ -64,18 +64,18 @@ public class Chef {
         Email = email;
     }
 
-    public int getLocationX() {
+    public double getLocationX() {
         return LocationX;
     }
 
-    public void setLocationX(int locationX) {
+    public void setLocationX(double locationX) {
         LocationX = locationX;
     }
-    public int getLocationY() {
+    public double getLocationY() {
         return LocationY;
     }
 
-    public void setLocationY(int locationY) {
+    public void setLocationY(double locationY) {
         LocationY = locationY;
     }
